+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - TCPDF_IMPORT class (it was not finished since 2013 and under development in 2013)
 - TCPDF_PARSER class (it was used by TCPDF_IMPORT)
 - TCPDF_FILTERS class (it was used by TCPDF_PARSER, and created in 2013, and never really changed)
+- Drop PHP 4 polyfill for `str_split` function in `include/barcodes/qrcode.php`
 
 ## [6.3.5] - 2020-02-14